print 64bits integer in hexadecimal

+++ include/stdlib.h
@@ -4,6 +4,6 @@
 int atoi(const char *str, int *dst);
 int htoi(const char *str, int *dst);
 char *itoa(char *tmpstr, long int arg);
-char *itoh(char *tmpstr, long int arg);
+char *itoh(char *tmpstr, unsigned long long arg);
 
 #endif
